Refactor FB table: Check for null in order to determine whether or not a variable contains data
Was using if (variable) which also fails if the variable evaluates to false.
if (variable != null) will detect whether or not a variable is set to null or its content is undefined. However, it will fail with an exception if the variable itself is undefined (this should never happen when processing the snapshot).